Card header guard clause
